// src/app/shared/models/lock-section-model.ts
export class LockSectionsModel {
  section1?: boolean = false;
  section2?: boolean = false;
  section3?: boolean = false;
  section4?: boolean = false;
  section5?: boolean = false;
  section6?: boolean = false;
  section7?: boolean = false;
  section8?: boolean = false;
  section9?: boolean = false;
  section10?: boolean = false;
  sectionFiles?: boolean = false;
}